docs: add PieChartWithCustomizedLabel example in TypeScript
